+++ plugins/azure-sites-common/src/types.ts
@@ -46,3 +46,11 @@ export type AzureSiteStartStopRequest = {
   resourceGroup: string;
   name: string;
 };
+
+/** @public */
+export type AzureSiteBackendRequest = {
+  subscription: string;
+  resourceGroup: string;
+  name: string;
+  entityRef: string;
+};