+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/fetch/plainFile.ts
@@ -33,7 +33,11 @@ export function createFetchPlainFileAction(options: {
 }) {
   const { reader, integrations } = options;
 
-  return createTemplateAction<{ url: string; targetPath: string }>({
+  return createTemplateAction<{
+    url: string;
+    targetPath: string;
+    token?: string;
+  }>({
     id: 'fetch:plain:file',
     description: 'Downloads single file and places it in the workspace.',
     examples,
@@ -52,6 +56,12 @@ export function createFetchPlainFileAction(options: {
             title: 'Target Path',
             description:
               'Target path within the working directory to download the file as.',
+            type: 'string',
+          },
+          token: {
+            title: 'Token',
+            description:
+              'An optional token to use for authentication when reading the resources.',
             type: 'string',
           },
         },
@@ -73,6 +83,7 @@ export function createFetchPlainFileAction(options: {
         baseUrl: ctx.templateInfo?.baseUrl,
         fetchUrl: ctx.input.url,
         outputPath,
+        token: ctx.input.token,
       });
     },
   });