+++ plugins/scaffolder-react/src/next/lib/schema.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 import { JsonObject } from '@backstage/types';
+import { stringify, parse } from 'flatted';
 import { FieldValidation, UiSchema } from '@rjsf/utils';
 
 function isObject(value: unknown): value is JsonObject {
@@ -57,9 +58,13 @@ function extractUiSchema(schema: JsonObject, uiSchema: JsonObject) {
       if (!isObject(schemaNode)) {
         continue;
       }
-      const innerUiSchema = {};
-      uiSchema[propName] = innerUiSchema;
-      extractUiSchema(schemaNode, innerUiSchema);
+
+      if (!isObject(uiSchema[propName])) {
+        const innerUiSchema = {};
+        uiSchema[propName] = innerUiSchema;
+      }
+
+      extractUiSchema(schemaNode, uiSchema[propName] as JsonObject);
     }
   }
 
@@ -123,7 +128,7 @@ export const extractSchemaFromStep = (
   inputStep: JsonObject,
 ): { uiSchema: UiSchema; schema: JsonObject } => {
   const uiSchema: UiSchema = {};
-  const returnSchema: JsonObject = JSON.parse(JSON.stringify(inputStep));
+  const returnSchema: JsonObject = parse(stringify(inputStep));
   extractUiSchema(returnSchema, uiSchema);
   return { uiSchema, schema: returnSchema };
 };