+++ plugins/git-release-manager/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-git-release-manager
 
+## 0.3.42-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.0-next.0
+  - @backstage/core-plugin-api@1.8.3-next.0
+  - @backstage/integration@1.9.0-next.0
+
 ## 0.3.41
 
 ### Patch Changes