+++ plugins/graphql-voyager/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-graphql-voyager
 
+## 0.1.12-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.0-next.0
+  - @backstage/core-plugin-api@1.8.3-next.0
+
 ## 0.1.11
 
 ### Patch Changes