+++ plugins/kubernetes/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-kubernetes
 
+## 0.11.5-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.0-next.0
+  - @backstage/catalog-model@1.4.4-next.0
+  - @backstage/plugin-kubernetes-react@0.3.0-next.1
+  - @backstage/core-plugin-api@1.8.3-next.0
+  - @backstage/plugin-catalog-react@1.9.4-next.1
+  - @backstage/plugin-kubernetes-common@0.7.4-next.1
+
 ## 0.11.5-next.0
 
 ### Patch Changes