+++ plugins/shortcuts/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-shortcuts
 
+## 0.3.19-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.0-next.0
+  - @backstage/core-plugin-api@1.8.3-next.0
+  - @backstage/theme@0.5.0
+  - @backstage/types@1.1.1
+
 ## 0.3.18
 
 ### Patch Changes